% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cluster.R
\docType{methods}
\name{cluster}
\alias{cluster}
\alias{cluster,concomp-method}
\title{cluster
cluster object and return class labels}
\usage{
cluster(obj, ...)

\S4method{cluster}{concomp}(obj, ...)
}
\arguments{
\item{obj}{object which can be clustered}

\item{obj}{a \code{concomp} object}

\item{exon}{a logical whether to return labels based on exon expression
(default = TRUE)}

\item{junc}{a logical whether to return labels based on junction expression
(default = TRUE)}

\item{both}{a logical whether to return labels based on joint clustering
(default = TRUE)}

\item{ej_w}{a numeric vector of length two specifying the relative sum of squares
for exon and junctions to be used if both = TRUE (default = c(1, 1))}

\item{agg_f}{a function used to aggregate the expression for each sample
to determine low vs. non-low expression (default = max)}

\item{min_cov}{a numeric value specifying the minimum threshold for agg_f of each
sample to determine low vs. non-low expression (default = 1)}

\item{min_n}{an integer value specifying the minimum number of non-low samples
       needed for clustering (default = 10)}
}
\value{
cluster labels as vector of cluster indices with same length
as number of observations
}
\description{
cluster
cluster object and return class labels

Method returns cluster labels for \code{concomp} object
}
\author{
Patrick Kimes
}
\keyword{internal}

