% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{concomp}
\alias{concomp}
\alias{concomp,GRangesList-method}
\alias{concomp,data.frame-method}
\alias{concomp-constructor}
\title{Connected Component (concomp) object}
\usage{
concomp(obj, ...)

\S4method{concomp}{GRangesList}(obj)

\S4method{concomp}{data.frame}(obj)
}
\arguments{
\item{obj}{a \code{GRangesList} with two elements named "e" and "j"}

\item{obj}{a \code{data.frame} containing all of the information}
}
\value{
\code{concomp} object
}
\description{
construct a connected component object from various input objects
}
\author{
Patrick Kimes
}

