% Generated by roxygen2 (4.0.2): do not edit by hand
\name{concomp2name}
\alias{concomp2name}
\title{match concomp with gene or transcript name}
\usage{
concomp2name(obj, txlist, txdb, orgdb = NULL)
}
\arguments{
\item{obj}{a \code{concomp} object}

\item{txlist}{a list of transcripts, e.g. \code{exonsBy(txdb)}}

\item{txdb}{a transcript database, e.g. \code{TxDb.Hsapiens.UCSC.hg19.knownGene}}

\item{orgdb}{a organism database, e.g. \code{org.Hs.eg.db} (default = NULL)}
}
\value{
a vector of transcript or gene names
}
\description{
Function matches concomp to nearest transcript and corresponding
gene names if both transcript list and organism db are provided. If
\code{orgdb} is not provided, the transcript IDs are returned rather than
the gene symbols
}
\author{
Patrick Kimes
}

