% Generated by roxygen2 (4.0.2): do not edit by hand
\name{diffsplice2name}
\alias{diffsplice2name}
\title{match matrix of diffsplice output with gene or transcript name}
\usage{
diffsplice2name(obj, ichr, seqlen, txlist, txdb = NULL, orgdb = NULL)
}
\arguments{
\item{obj}{a \code{data.frame} loaded from \code{\link{readchr}}}

\item{ichr}{a numeric value specifying the chromosome number, e.g. 9}

\item{seqlen}{a numeric value specifying the sequence length corresponding to the
chromosome specified in \code{ichr}, e.g. take from \code{seqlen(txdb)}}

\item{txlist}{a list of transcripts, e.g. \code{exonsBy(txdb)}}

\item{txdb}{a transcript database, e.g. \code{TxDb.Hsapiens.UCSC.hg19.knownGene}
(default = NULL)}

\item{orgdb}{a organism database, e.g. \code{org.Hs.eg.db} (default = NULL)}
}
\value{
a \code{data.frame} with connected component IDs and overlapping
        transcript or gene names
}
\description{
Function matches \code{data.frame} of diffsplice output to nearest
transcripts using \code{txdb} and \code{txlist}. This is further mapped to
gene names if \code{orgdb} is specified. If \code{orgdb} is not provided,
transcript IDs are returned instead of gene symbols.
}
\author{
Patrick Kimes
}

