% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{eCoverage<-}
\alias{eCoverage<-}
\title{eCoverage replace
replace \code{eCoverage} slot}
\usage{
eCoverage(obj, ...) <- value
}
\arguments{
\item{obj}{object with slot \code{eCoverage}}
}
\description{
eCoverage replace
replace \code{eCoverage} slot
}
\keyword{internal}

