% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{eCoverage}
\alias{eCoverage}
\title{eCoverage
return \code{eCoverage} slot}
\usage{
eCoverage(obj)
}
\arguments{
\item{obj}{object with slot \code{eCoverage}}
}
\description{
eCoverage
return \code{eCoverage} slot
}
\keyword{internal}

