% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{eRanges<-}
\alias{eRanges<-}
\title{eRanges replace
replace \code{eRanges} slot}
\usage{
eRanges(obj, ...) <- value
}
\arguments{
\item{obj}{object with slot \code{eRanges}}
}
\description{
eRanges replace
replace \code{eRanges} slot
}
\keyword{internal}

