% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{eRanges}
\alias{eRanges}
\title{eRanges
return \code{eRanges} slot}
\usage{
eRanges(obj)
}
\arguments{
\item{obj}{object with slot \code{eRanges}}
}
\description{
eRanges
return \code{eRanges} slot
}
\keyword{internal}

