% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{exons<-}
\alias{exons<-}
\title{exons replace
replace \code{exons} slot}
\usage{
exons(obj, ...) <- value
}
\arguments{
\item{obj}{object with slot \code{exons}}
}
\description{
exons replace
replace \code{exons} slot
}
\keyword{internal}

