% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{exons}
\alias{exons}
\title{exons
return \code{exons} slot}
\usage{
exons(obj)
}
\arguments{
\item{obj}{object with slot \code{exons}}
}
\description{
exons
return \code{exons} slot
}
\keyword{internal}

