% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{exonValues<-}
\alias{exonValues<-}
\title{exonValues replace
replace \code{exonValues} slot}
\usage{
exonValues(obj, ...) <- value
}
\arguments{
\item{obj}{object with slot \code{exonValues}}
}
\description{
exonValues replace
replace \code{exonValues} slot
}
\keyword{internal}

