% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{exonValues}
\alias{exonValues}
\title{exonValues
return \code{exonValues} slot}
\usage{
exonValues(obj)
}
\arguments{
\item{obj}{object with slot \code{exonValues}}
}
\description{
exonValues
return \code{exonValues} slot
}
\keyword{internal}

