% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{jCoverage<-}
\alias{jCoverage<-}
\title{jCoverage replace
replace \code{jCoverage} slot}
\usage{
jCoverage(obj, ...) <- value
}
\arguments{
\item{obj}{object with slot \code{jCoverage}}
}
\description{
jCoverage replace
replace \code{jCoverage} slot
}
\keyword{internal}

