% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{jCoverage}
\alias{jCoverage}
\title{jCoverage
return \code{jCoverage} slot}
\usage{
jCoverage(obj)
}
\arguments{
\item{obj}{object with slot \code{jCoverage}}
}
\description{
jCoverage
return \code{jCoverage} slot
}
\keyword{internal}

