% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{jRanges}
\alias{jRanges}
\title{jRanges
return \code{jRanges} slot}
\usage{
jRanges(obj)
}
\arguments{
\item{obj}{object with slot \code{jRanges}}
}
\description{
jRanges
return \code{jRanges} slot
}
\keyword{internal}

