% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{juncs}
\alias{juncs}
\title{juncs
return \code{juncs} slot}
\usage{
juncs(obj)
}
\arguments{
\item{obj}{object with slot \code{juncs}}
}
\description{
juncs
return \code{juncs} slot
}
\keyword{internal}

