% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{juncValues<-}
\alias{juncValues<-}
\title{juncValues replace
replace \code{juncValues} slot}
\usage{
juncValues(obj, ...) <- value
}
\arguments{
\item{obj}{object with slot \code{juncValues}}
}
\description{
juncValues replace
replace \code{juncValues} slot
}
\keyword{internal}

