% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_klk12_hnsc_cc.R
\docType{data}
\name{klk12_hnsc_cc}
\alias{klk12_hnsc_cc}
\title{connected component of KLK12 Expression in HNSC}
\format{a concomp object}
\source{
The Cancer Genome Atlas Project
}
\usage{
data(klk12_hnsc_cc)
}
\description{
concomp object of exon and splice junction expression values
for a cohort of 279 head and neck squamous cell (HNSC) samples
with 16 exons and 16 junctions.
}
\author{
Patrick Kimes
}

