% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_klk12_lusc_gr.R
\docType{data}
\name{klk12_lusc_gr}
\alias{klk12_lusc_gr}
\title{GRanges of KLK12 Expression in LUSC}
\format{a GRanges object}
\source{
The Cancer Genome Atlas Project
}
\usage{
data(klk12_lusc_gr)
}
\description{
GRanges object of exon and splice junction boundaries with metadata columns
providing exon/junction labels and expression values for a cohort of 177
lung sequamous cell (LUSC) samples with 13 exons and 12 junctions.
}
\author{
Patrick Kimes
}

