% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{long_letters}
\alias{long_letters}
\title{function defining useful vector of letters}
\usage{
long_letters()
}
\description{
function defining useful vector of letters
}
\author{
Patrick Kimes
}
\keyword{internal}

