% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{name2gidx}
\alias{name2gidx}
\title{match gene name (symbol) to connected component index}
\usage{
name2gidx(name, obj, txlist, txdb, orgdb)
}
\arguments{
\item{name}{a \code{character} string specifying the gene of interest}

\item{obj}{a \code{GRangesList} of connected component ranges, each name corresponding
to the 'gene index'}

\item{txlist}{a list of transcripts, e.g. \code{exonsBy(txdb)}}

\item{txdb}{a transcript database, e.g. \code{TxDb.Hsapiens.UCSC.hg19.knownGene}
(default = NULL)}

\item{orgdb}{a organism database, e.g. \code{org.Hs.eg.db} (default = NULL)}
}
\value{
a \code{data.frame} with connected component IDs and overlapping
        transcript or gene names
}
\description{
Function matches a character string gene symbol to overlapping connected components
according to annotations in \code{txdb} and \code{txlist}. Note to self: R follows
copy-on-write semantics.
}
\author{
Patrick Kimes
}

