% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{pseudoArc}
\alias{.pseudoArc}
\alias{pseudoArc}
\title{helper function for adding arcs to ggplot2 object via geom_path()}
\usage{
.pseudoArc(xmin = 0, xmax = 10, ymin = 0, height = 10, npoints = 100)
}
\arguments{
\item{xmin}{numeric value of left end of arc}

\item{xmax}{numeric value of right end of arc}

\item{ymin}{numeric value of bottom of arc}

\item{height}{numeric value of arc height}

\item{npoints}{integer number of points to draw arc}
}
\description{
modified from: http://stackoverflow.com/questions/6862742/
}
\author{
Patrick Kimes
}
\keyword{internal}

