% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{rgb2hex}
\alias{.rgb2hex}
\alias{rgb2hex}
\title{helper for converting rgb triple to hex color}
\usage{
.rgb2hex(rgb)
}
\arguments{
\item{rgb}{matrix with three columns with RGB values in each row}
}
\value{
vector of hex colors with length same as number of rows in
        input matrix
}
\description{
helper for converting rgb triple to hex color
}
\author{
Patrick Kimes
}
\keyword{internal}

