% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicegrahm-helpers.R
\name{sg_create}
\alias{sg_create}
\title{construct ggplot2 object for plotting}
\usage{
sg_create(gr_e, gr_j, vals_e, vals_j, j_incl, log_base, log_shift, bin, n, p_j,
  y_flip = FALSE, same_scale_n = n)
}
\arguments{
\item{gr_e}{\code{GenomicRanges} for exons}

\item{gr_j}{\code{GenomicRanges} for junctions}

\item{vals_e}{matrix of exon coverages}

\item{vals_j}{matrix of junction coverages}

\item{j_incl}{see \code{splicegrahm} documentation}

\item{log_base}{see \code{splicegrahm} documentation}

\item{log_shift}{see \code{splicegrahm} documentation}

\item{bin}{see \code{splicegrahm} documentation}

\item{n}{number of samples in \code{vals_e}, \code{vals_j}}

\item{p_j}{number of junctions in \code{gr_j}}

\item{y_flip}{logical whether model should be flipped on
vertical axis (default = FALSE)}

\item{same_scale_n}{number of samples that should be used to set vertical scaling of
\code{splicegrahm2} plot (if \code{splicegrahm2} parameter
\code{same_scale = FALSE}, then simply \code{n}) (default = n)}
}
\description{
construct ggplot2 object for plotting
}
\author{
Patrick Kimes
}
\keyword{internal}

