% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splicegrahm-helpers.R
\name{sg_drawjuncs}
\alias{sg_drawjuncs}
\title{add junction annotations to base ggplot2 object for splicegrahm plot}
\usage{
sg_drawjuncs(sg_obj, sg_df, j_incl, use_blk, iflip, gr_e, gr_j, vals_j, n, p_j,
  highlight, mirror = FALSE, same_scale_n = n)
}
\arguments{
\item{sg_obj}{\code{ggplot} object created by \code{sg_drawbase}}

\item{sg_df}{data.frame output from \code{sg_create}}

\item{j_incl}{see \code{splicegrahm} documentation}

\item{use_blk}{see \code{splicegrahm} documentation}

\item{iflip}{logical values whether the plotting should be reversed}

\item{gr_e}{\code{GenomicRanges} for exons}

\item{gr_j}{\code{GenomicRanges} for junctions}

\item{vals_j}{matrix of junction coverages}

\item{n}{number of samples in \code{vals_e}, \code{vals_j}}

\item{p_j}{number of junctions in \code{gr_j}}

\item{highlight}{see \code{splicegrahm} documentation}

\item{mirror}{logical whether model should be flipped on
vertical axis (defualt = FALSE)}

\item{same_scale_n}{number of samples that should be used to set vertical scaling of
       \code{splicegrahm2} plot (if \code{splicegrahm2} parameter
       \code{same_scale = FALSE}, then simply \code{n}) (default = n)}
}
\description{
add junction annotations to base ggplot2 object for splicegrahm plot
}
\author{
Patrick Kimes
}
\keyword{internal}

