% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concomp-show.R
\docType{methods}
\name{show-concomp}
\alias{show,concomp-method}
\alias{show-concomp}
\title{show summary of concomp object}
\usage{
\S4method{show}{concomp}(object)
}
\arguments{
\item{object}{a \code{concomp} object}
}
\value{
Descrption of \code{concomp} object
}
\description{
show quick summary of \code{concomp} object
}
\author{
Patrick Kimes
}
\keyword{internal}

