% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splicegralp-helpers.R
\name{sl_create}
\alias{sl_create}
\title{construct ggplot2 object for plotting}
\usage{
sl_create(gr_e, e_loads)
}
\arguments{
\item{gr_e}{\code{GenomicRanges} for exons}

\item{e_loads}{see \code{splicegralp} documentation}
}
\description{
construct ggplot2 object for plotting
}
\author{
Patrick Kimes
}
\keyword{internal}

