% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splicegralp-helpers.R
\name{sl_drawbase}
\alias{sl_drawbase}
\title{construct base ggplot2 object for splicegralp plot}
\usage{
sl_drawbase(sl_df, gr_e, gr_j, p_e, p_j, e_loads, j_loads, genomic, use_blk,
  load_lims, n_vec)
}
\arguments{
\item{sl_df}{data.frame output from \code{sl_create}}

\item{gr_e}{\code{GenomicRanges} for exons}

\item{gr_j}{\code{GenomicRanges} for junctions}

\item{p_e}{number of exons in \code{gr_e}}

\item{genomic}{see \code{splicegralp} documentation}

\item{use_blk}{see \code{splicegralp} documentation}

\item{p_e}{number of junctions in \code{gr_j}}

\item{bin}{see \code{splicegralp} documentation}
}
\description{
construct base ggplot2 object for splicegralp plot
}
\author{
Patrick Kimes
}
\keyword{internal}

