% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/slots-generic.R
\name{eRanges}
\alias{eCoverage}
\alias{eCoverage<-}
\alias{eRanges}
\alias{eRanges<-}
\alias{jCoverage}
\alias{jCoverage<-}
\alias{jRanges}
\alias{jRanges<-}
\alias{slots-generic}
\title{Generic Methods for S4 Class Slots in spliceclust}
\usage{
eRanges(object)

jRanges(object)

eCoverage(object)

jCoverage(object)

eRanges(object, ...) <- value

jRanges(object, ...) <- value

eCoverage(object, ...) <- value

jCoverage(object, ...) <- value
}
\arguments{
\item{object}{object with corresponding slot}

\item{...}{other parameters}

\item{value}{new value for lsot}
}
\description{
Getters can be used to access slot values, and setters can be used
for specifying new slot values.
}
\author{
Patrick Kimes
}

