% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splicepcp-helpers.R
\name{sp_create}
\alias{sp_create}
\title{construct ggplot2 object for plotting}
\usage{
sp_create(gr_e, gr_j, vals_e, vals_j, j_incl, log_base, log_shift, bin, n, p_e,
  p_j)
}
\arguments{
\item{gr_e}{\code{GenomicRanges} for exons}

\item{gr_j}{\code{GenomicRanges} for junctions}

\item{vals_e}{matrix of exon coverages}

\item{vals_j}{matrix of junction coverages}

\item{j_incl}{see \code{splicegrahm} documentation}

\item{log_base}{see \code{splicegrahm} documentation}

\item{log_shift}{see \code{splicegrahm} documentation}

\item{bin}{see \code{splicegrahm} documentation}

\item{n}{number of samples in \code{vals_e}, \code{vals_j}}

\item{p_j}{number of junctions in \code{gr_j}}
}
\description{
construct ggplot2 object for plotting
}
\author{
Patrick Kimes
}
\keyword{internal}

