% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicepcp-helpers.R
\name{sp_drawbase}
\alias{sp_drawbase}
\title{construct base ggplot2 object for splicegrahm plot}
\usage{
sp_drawbase(sp_df, highlight, genomic, log_base, bin, gr_e, p_e)
}
\arguments{
\item{sp_df}{data.frame output from \code{sp_create}}

\item{highlight}{see \code{splicepcp} documentation}

\item{genomic}{see \code{splicepcp} documentation}

\item{log_base}{see \code{splicepcp} documentation}

\item{bin}{see \code{splicepcp} documentation}

\item{gr_e}{\code{GenomicRanges} for exons}

\item{p_e}{number of exons in \code{gr_e}}
}
\description{
construct base ggplot2 object for splicegrahm plot
}
\author{
Patrick Kimes
}
\keyword{internal}

