% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splicepcp-helpers.R
\name{sp_drawmodel}
\alias{sp_drawmodel}
\title{construct ggplot object with connected component model for splicepcp}
\usage{
sp_drawmodel(sp_df, gr_e, gr_j, p_j, genomic)
}
\arguments{
\item{sp_df}{data.frame output from \code{sg_create}}

\item{gr_e}{\code{GenomicRanges} for exons}

\item{gr_j}{\code{GenomicRanges} for junctions}

\item{p_j}{number of junctions in \code{gr_j}}

\item{genomic}{see \code{splicepcp} documentation}
}
\description{
construct ggplot object with connected component model for splicepcp
}
\author{
Patrick Kimes
}
\keyword{internal}

