% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spliceclust-package.R
\docType{package}
\name{spliceclust-package}
\alias{spliceclust}
\alias{spliceclust-package}
\title{spliceclust}
\description{
This package includes functions for analyzing and vizualizing
coverage/expression along exons and junction of a single locus
across multiple samples/experiments. In this package, we use
"connected component" to refer to a collection of exons
and splice junctions which correspond roughly to individual
gene.
}
\author{
Patrick Kimes \email{patrick.kimes@gmail.com}
}
\keyword{package}

