% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicegrahm.R
\docType{methods}
\name{splicegrahm-concomp}
\alias{splicegrahm,concomp-method}
\alias{splicegrahm-concomp}
\title{splicegrahm method}
\usage{
\S4method{splicegrahm}{concomp}(obj, ...)
}
\description{
\code{splicegrahm} method for \code{concomp} class object.
See \code{splicegrahm} documentation for more details.
}
\seealso{
splicegrahm
}
\keyword{internal}

