% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splicegrahm2.R
\docType{methods}
\name{splicegrahm2-concomp}
\alias{splicegrahm2,concomp-method}
\alias{splicegrahm2-concomp}
\title{splicegrahm2 method}
\usage{
\S4method{splicegrahm2}{concomp,concomp}(obj1, obj2 = "concomp", ...)
}
\description{
splicegrahm2 method
}
\keyword{internal}

