% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{splicegrahm2-generic}
\alias{splicegrahm2-generic}
\title{Two-class Splice Graph HeatMap (SpliceGraHM2)}
\usage{
splicegrahm2(obj1, obj2, ...)
}
\description{
Method for plotting both the exon and junction values of a single
connected component for two separate clusters or populations
}
\keyword{internal}

