% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicegralp.R
\docType{methods}
\name{splicegralp-concomp}
\alias{splicegralp,concomp-method}
\alias{splicegralp-concomp}
\title{splicegralp method}
\usage{
\S4method{splicegralp}{concomp}(obj, ...)
}
\description{
\code{splicegralp} method for \code{concomp} class object.
See \code{splicegralp} documentation for more details.
}
\seealso{
splicegralp
}
\keyword{internal}

