% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{splicegralp}
\alias{splicegralp}
\alias{splicegralp,concomp-method}
\title{SpliceGraLP: splice graph loadings plot}
\usage{
\S4method{splicegralp}{concomp}(obj, ...)
}
\arguments{
\item{obj}{a \code{concomp} object with exon and junction information}

\item{...}{other parameters to be passed}

\item{e_loads}{a matrix of exon loadings in columns}

\item{j_loads}{a matrix of junction loadings in columns (default = NULL)}

\item{load_lims}{a vector of length 2 that should be used as the bounds on
the color scale, e.g. \code{c(-1, 1)} if plotting PCA loadings}

\item{genomic}{a logical whether genomic coordinates should be used to
plot the heatmap (default = TRUE)}

\item{ex_use}{a numeric specifying the proportion of the plot exons should occupy if
non-genomic coordinate plotting is desired (default = 2/3)}

\item{flip_neg}{a logical whether to flip plotting of genes on negative strand
to run left to right (default = TRUE)}

\item{use_blk}{a logical whether to use a black background (default = FALSE)}

\item{txlist}{a GRangesList of transcripts or genes which should be queried and
added to the plot if falling within the region of the connected component
(default = NULL)}

\item{txlist}{a GRangesList of transcripts or genes which should be queried and
added to the plot if falling within the region of the connected component
(default = NULL)}

\item{txdb}{a transcript database which can be used to query the transcript IDs
identified from txlist (default = NULL)}

\item{orgdb}{a database that can be queried using keys obtained from \code{txdb}
to determine corresponding gene symbols (default = NULL)}
}
\value{
a ggplot2 plot showing the specified number of loadings
}
\description{
Splice graphs with loadings shown along the exons and junctions to illustrate
expression patterns across a connected component. The function can be used to
illustrate, e.g. differences across groups/clusters or principal component
loadings.
}
\author{
Patrick Kimes
}

