% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{splicepca-generic}
\alias{splicepca-generic}
\title{Splice Graph PCA Plot}
\usage{
splicepca(obj, ...)
}
\description{
Method for plotting PCA loading vectors for a \code{concomp}
object (wrapper to splicegralp)
}
\keyword{internal}

