% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{splicepcp-generic}
\alias{splicepcp-generic}
\title{Splice Graph Parallel Coordinates Plot}
\usage{
splicepcp(obj, ...)
}
\description{
Method for plotting exon expression for a \code{concomp}
object with each exon given equal width with heights corresponding
to expression level
}
\keyword{internal}

