% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splicepcp.R
\docType{methods}
\name{splicepcp}
\alias{splicepcp}
\alias{splicepcp,concomp-method}
\title{SplicePCP: splice graph parallel coordinates plot}
\usage{
\S4method{splicepcp}{concomp}(obj, ...)
}
\arguments{
\item{obj}{a \code{concomp} object}

\item{...}{other parameters to be passed}

\item{log_base}{a numeric specifying the scaling of expression values at each exon,
which 0 resulting in no long scaling being applied (default = 10)}

\item{log_shift}{a numeric specifying the shift to be used in the log transformation
for handling 0 values (default = 1)}

\item{genomic}{a logical whether genomic coordinates should be used to
plot the heatmap (default = TRUE)}

\item{ex_use}{a numeric specifying the proportion of the plot exons should occupy if
non-genomic coordinate plotting is desired (default = 2/3)}

\item{flip_neg}{a logical whether to flip plotting of genes on negative strand
to run left to right (default = TRUE)}

\item{highlight}{a vector of labels to highlight samples in groups or clusters
(default = NULL)}

\item{imodel}{a logical whether to include the connected component plot generated
by splicegralp or splicegrahm to the plot (default = TRUE)}

\item{j_incl}{a logical whether to include the junction information as a
parallel coordinates track (default = FALSE)}

\item{eps}{a numeric value specifying the number of base pairs around \code{obj} to look
for overlapping gene models, if eps = NULL, then all overlapping gene models are
included (default = 1e4)}

\item{txlist}{a GRangesList of transcripts or genes which should be queried and
added to the plot if falling within the region of the connected component
(default = NULL)}

\item{txdb}{a transcript database which can be used to query the transcript IDs
identified from txlist (default = NULL)}

\item{orgdb}{a database that can be queried using keys obtained from \code{txdb}
to determine corresponding gene symbols (default = NULL)}
}
\value{
a ggplot2 plot showing the splice graph with heatmaps shown at each node
and each splice
}
\description{
Splice graph expression plot with exons along horizontal axis with vertical
axis showing expression level. The plot provides an alternative to the
heatmaps used to visualize expression using the \code{splicegrahm} and
\code{splicegralp} functions.

splicepcp method
}
\author{
Patrick Kimes
}
\keyword{internal}

