% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R, R/concomp-class.R, R/concomp-constructor.R, R/concomp-getters.R, R/concomp-setters.R, R/concomp-show.R
\docType{class}
\name{concomp-class}
\alias{concomp}
\alias{concomp,GRangesList-method}
\alias{concomp,data.frame-method}
\alias{concomp-class}
\alias{eCoverage,concomp-method}
\alias{eCoverage<-,concomp-method}
\alias{eRanges,concomp-method}
\alias{eRanges<-,concomp-method}
\alias{jCoverage,concomp-method}
\alias{jCoverage<-,concomp-method}
\alias{jRanges,concomp-method}
\alias{jRanges<-,concomp-method}
\alias{show,concomp-method}
\title{Connected Component}
\usage{
concomp(obj, ...)

\S4method{concomp}{GRangesList}(obj)

\S4method{concomp}{data.frame}(obj)

\S4method{eRanges}{concomp}(obj)

\S4method{jRanges}{concomp}(obj)

\S4method{eCoverage}{concomp}(obj)

\S4method{jCoverage}{concomp}(obj)

\S4method{eRanges}{concomp}(obj) <- value

\S4method{jRanges}{concomp}(obj) <- value

\S4method{eCoverage}{concomp}(obj) <- value

\S4method{jCoverage}{concomp}(obj) <- value

\S4method{show}{concomp}(object)
}
\arguments{
\item{obj}{a \code{GRangesList} with two \code{GRanges} named "e" and "j"
corresponding to exon and junction ranges and coverage for a
collection of samples to be used to construct a \code{concomp} object.
Alternative, can be a \code{data.frame} with gene model columns:
'chr', 'seqlengths', 'gStart', 'gStop', and 'kind' (e.g. 'e' or 'j'), 
as well as per-sample coverage columns: 's1', 's2', etc.}

\item{...}{other parameters to be passed to constructor}

\item{value}{new value for \code{concomp} slot}

\item{object}{a \code{concomp} object for show method}

\item{obj}{a \code{concomp} object for getters and setters}
}
\description{
S4 class encapsulating exon and junction information for single
connected component
}
\details{
The columns of \code{eCoverage} and \code{jCoverage} correspond
to individual samples or experiments, and should be of equal length.
}
\section{Methods (by generic)}{
\itemize{
\item \code{eRanges}: get \code{eRanges} slot

\item \code{jRanges}: get \code{jRanges} slot

\item \code{eCoverage}: get \code{eCoverage} slot

\item \code{jCoverage}: get \code{jCoverage} slot

\item \code{eRanges<-}: set \code{eRanges} slot

\item \code{jRanges<-}: set \code{jRanges} slot

\item \code{eCoverage<-}: set \code{eCoverage} slot

\item \code{jCoverage<-}: set \code{jCoverage} slot
}}
\section{Slots}{

\describe{
\item{\code{eRanges}}{a \code{GRanges} object containing the exonic information}

\item{\code{jRanges}}{a \code{GRanges} object containing the junction information}

\item{\code{eCoverage}}{a matrix of exon coverage values, e.g. RPKM, with
number of rows equal to the length of \code{exons}, or \code{NULL}
if not available}

\item{\code{jCoverage}}{a matrix of junction coverage values, e.g. number of
spanning reads, with number of rows equal to the length of
\code{juncs}, or \code{NULL} if not available}
}}
\author{
Patrick Kimes
}

