% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/concomp-getters.R, R/concomp-setters.R
\docType{class}
\name{concomp-class}
\alias{concomp-class}
\alias{exonValues,concomp-method}
\alias{exonValues<-,concomp-method}
\alias{exons,concomp-method}
\alias{exons<-,concomp-method}
\alias{juncValues,concomp-method}
\alias{juncValues<-,concomp-method}
\alias{juncs,concomp-method}
\alias{juncs<-,concomp-method}
\title{connected component object}
\usage{
\S4method{exons}{concomp}(obj)

\S4method{juncs}{concomp}(obj)

\S4method{exonValues}{concomp}(obj)

\S4method{juncValues}{concomp}(obj)

\S4method{exons}{concomp}(obj, ...) <- value

\S4method{juncs}{concomp}(obj, ...) <- value

\S4method{exonValues}{concomp}(obj, ...) <- value

\S4method{juncValues}{concomp}(obj, ...) <- value
}
\arguments{
\item{obj}{a \code{concomp} object}
}
\description{
S4 class encapsulating exon and junction
information for single connected component
}
\details{
The columns of \code{exonValues} and \code{juncValues} correspond
to individual samples or experiments, and should be of equal length.
}
\section{Methods (by generic)}{
\itemize{
\item \code{exons}: get \code{exons} slot

\item \code{juncs}: get \code{juncs} slot

\item \code{exonValues}: get \code{exonValues} slot

\item \code{juncValues}: get \code{juncValues} slot

\item \code{exons<-}: set \code{exons} slot

\item \code{juncs<-}: set \code{juncs} slot

\item \code{exonValues<-}: set \code{exonValues} slot

\item \code{juncValues<-}: set \code{juncValues} slot
}}
\section{Slots}{

\describe{
\item{\code{exons}}{a \code{GRanges} object containing the exonic information}

\item{\code{juncs}}{a \code{GRanges} object containing the junction information}

\item{\code{exonValues}}{a matrix of exon coverage values, e.g. RPKM, with
number of rows equal to the length of \code{exons}, or \code{NULL}
if not available}

\item{\code{juncValues}}{a matrix of junction coverage values, e.g. number of
spanning reads, with number of rows equal to the length of
\code{juncs}, or \code{NULL} if not available}
}}
\author{
Patrick Kimes
}

