% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{exonValues}
\alias{exonValues}
\title{exonValues
return \code{exonValues} slot}
\usage{
exonValues(obj)
}
\arguments{
\item{obj}{object with slot \code{exonValues}}
}
\description{
exonValues
return \code{exonValues} slot
}
\keyword{internal}

