% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{juncValues}
\alias{juncValues}
\title{juncValues
return \code{juncValues} slot}
\usage{
juncValues(obj)
}
\arguments{
\item{obj}{object with slot \code{juncValues}}
}
\description{
juncValues
return \code{juncValues} slot
}
\keyword{internal}

