% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readchr.R
\name{readchr}
\alias{readchr}
\title{read in data for single chromosome}
\usage{
readchr(file, n)
}
\arguments{
\item{file}{character string to data directory}

\item{n}{integer value specifying number of samples}
}
\value{
a \code{data.frame} object containing splicing information in
\code{6 + n} columns. The columns are:
\code{gIdx}, \code{gStart}, \code{gStop}, \code{kind}, \code{start},
\code{stop}, \code{s1}, \code{s2}, ...
}
\description{
Function to read in data for single chromosome output from
\code{diffsplice_bam} function provided by Yin Hu.
}
\author{
Patrick Kimes
}

