% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot-helpers.R
\name{sampl_sort}
\alias{sampl_sort}
\title{function defining all of the sort orders}
\usage{
sampl_sort(sort_idx, vals_e, vals_j)
}
\arguments{
\item{sort_idx}{integer value, see documentation for \code{splicegrahm}}

\item{vals_e}{matrix of exon coverages}

\item{vals_j}{matrix of junction coverages}
}
\description{
function defining all of the sort orders
}
\author{
Patrick Kimes
}
\keyword{internal}

