% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splicegrahm-helpers.R
\name{sg_drawbase}
\alias{sg_drawbase}
\title{construct base ggplot2 object for splicegrahm plot}
\usage{
sg_drawbase(sg_df, use_blk, j_incl, genomic, gr_e, log_base, bin, n, highlight,
  p_j, iflip, y_flip = FALSE, jointplot = FALSE)
}
\arguments{
\item{sg_df}{data.frame output from \code{sg_create}}

\item{use_blk}{see \code{splicegrahm} documentation}

\item{j_incl}{see \code{splicegrahm} documentation}

\item{genomic}{see \code{splicegrahm} documentation}

\item{gr_e}{\code{GenomicRanges} for exons}

\item{log_base}{see \code{splicegrahm} documentation}

\item{bin}{see \code{splicegrahm} documentation}

\item{n}{number of samples in \code{vals_e}, \code{vals_j}}

\item{highlight}{see \code{splicegrahm} documentation}

\item{p_j}{number of junctions in \code{gr_j}}

\item{iflip}{logical whether model will be on negative strand}

\item{y_flip}{logical whether model should be flipped on
vertical axis (defualt = FALSE)}

\item{jointplot}{logical whether to exclude setting of y_axis
       parameters in case of joint plotting with splicegrahm2
       (default = FALSE)}
}
\description{
construct base ggplot2 object for splicegrahm plot
}
\author{
Patrick Kimes
}
\keyword{internal}

