% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splicegrahm.R
\docType{methods}
\name{splicegrahm-concomp}
\alias{splicegrahm,concomp-method}
\alias{splicegrahm-concomp}
\title{splicegrahm method}
\usage{
\S4method{splicegrahm}{concomp}(obj, ...)
}
\description{
splicegrahm method
}
\keyword{internal}

