% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\docType{methods}
\name{splicegrahm-generic}
\alias{splicegrahm-generic}
\title{Splice Graph HeatMap (SpliceGraHM)}
\usage{
splicegrahm(obj, ...)
}
\description{
Method for plotting both the exon and junction values of a single
connected component
}
\keyword{internal}

