% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splicegrahm.R
\name{splicegrahm}
\alias{splicegrahm}
\title{SpliceGraHM: splice graph heatmap}
\arguments{
\item{obj}{a \code{concomp} object}

\item{sort_sep}{a logical whether to sort each exon, junction separately
(default = FALSE)}

\item{sort_idx}{an integer value specifying the order of the samples in
each exon, see details for more information on all possible
input, if length is n, then this ordering is used (default = 1)}

\item{log_base}{a numeric specifying the scale of the binning for the
plotting of expression values at each exon, which 0 resulting in no long
scaling being applied (default = 10)}

\item{log_shift}{a numeric specifying the shift to be used in the log transformation
for handling 0 values (default = 1)}

\item{bin}{a logical whether to bin the values for easier viewing (default = TRUE)}

\item{genomic}{a logical whether genomic coordinates should be used to
plot the heatmap (default = TRUE)}

\item{ex_use}{a numeric specifying the proportion of the plot exons should occupy if
non-genomic coordinate plotting is desired (default = 2/3)}

\item{flip_neg}{a logical whether to flip plotting of genes on negative strand
to run left to right (default = TRUE)}

\item{j_incl}{a logical whether to include heatmaps for junctions
(default = FALSE)}

\item{highlight}{a vector of labels to highlight samples in groups or clusters
(default = NULL)}

\item{use_blk}{a logical whether to use a black background (default = FALSE)}

\item{eps}{a numeric value specifying the number of base pairs around \code{obj} to look
for overlapping gene models, if eps = NULL, then all overlapping gene models are
included (default = 1e4)}

\item{txlist}{a GRangesList of transcripts or genes which should be queried and
added to the plot if falling within the region of the connected component
(default = NULL)}

\item{txdb}{a transcript database which can be used to query the transcript IDs
identified from txlist (default = NULL)}

\item{orgdb}{a database that can be queried using keys obtained from \code{txdb}
to determine corresponding gene symbols (default = NULL)}

\item{title}{a character string title printed at the top of plot (default = "")}

\item{...}{other parameters to be passed}
}
\value{
a ggplot2 plot showing the splice graph with heatmaps shown at each node
and each splice
}
\description{
Splice graphs as disjoint heatmaps for the comparative visualization
of splice graphs across multiple samples in a single figure.
}
\details{
sort_idx can take values of either:
\itemize{
\item{\code{1}}: sort on first exon
\item{\code{2}}: sort on PC 2
\item{\code{3}}: sort on mean exon coverage
\item{\code{4}}: sort on mean exon log-coverage
\item{\code{5}}: rev sort on mean exon coverage
\item{\code{6}}: rev sort on mean exon log-coverage
}
}
\author{
Patrick Kimes
}

