% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splicegrahm2.R
\docType{methods}
\name{splicegrahm2-concomp}
\alias{splicegrahm2,concomp,concomp-method}
\alias{splicegrahm2-concomp}
\title{splicegrahm2 method}
\usage{
\S4method{splicegrahm2}{concomp,concomp}(obj1, obj2, ...)
}
\description{
\code{splicegrahm2} method for \code{concomp} class object.
See \code{splicegrahm2} documentation for more details.
}
\seealso{
splicegrahm2
}
\keyword{internal}

