% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splicegralp.R
\docType{methods}
\name{splicegralp-concomp}
\alias{splicegralp,concomp-method}
\alias{splicegralp-concomp}
\title{splicegralp method}
\usage{
\S4method{splicegralp}{concomp}(obj, ...)
}
\description{
splicegralp method
}
\keyword{internal}

