% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{splicegralp-generic}
\alias{splicegralp-generic}
\title{Splice Graph Loadings Plot (SpliceGraLP)}
\usage{
splicegralp(obj, ...)
}
\description{
Method for plotting different loading vectors for a \code{concomp}
object, e.g. principal component directions
}
\keyword{internal}

