% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{splicePCA-generic}
\alias{splicePCA-generic}
\title{Splice Graph PCA Plot}
\usage{
splicePCA(obj, ...)
}
\description{
Method for plotting PCA loading vectors for a \code{concomp}
object (wrapper to splicegralp)
}
\keyword{internal}

